# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyTrip - Road Trip Planner API",
...
    "version": "1.0.0",
    "contact": {
      "name": "MyTrip API Support",
      "email": "support@mytrip.com",
      "url": "https://github.com/AdarBahar/MyTrip"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/places/v1/places/suggest": {
      "get": {
        "tags": [
          "places",
          "places-typeahead",
          "places-typeahead"
        ],
        "summary": "Type-ahead Suggestions",
        "description": "Get fast, lightweight place suggestions while typing",
        "operationId": "get_suggestions_places_v1_places_suggest_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "description": "Search query (minimum 1 character)",
              "title": "Q"
            },
            "description": "Search query (minimum 1 character)",
            "example": "montef"
          },
          {
            "name": "session_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Session token for grouping requests",
              "title": "Session Token"
            },
            "description": "Session token for grouping requests",
            "example": "st_abc123"
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "User latitude for proximity bias",
              "title": "Lat"
            },
            "description": "User latitude for proximity bias",
            "example": 32.07
          },
          {
            "name": "lng",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "User longitude for proximity bias",
              "title": "Lng"
            },
            "description": "User longitude for proximity bias",
            "example": 34.78
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 50000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bias radius in meters",
              "title": "Radius"
            },
            "description": "Bias radius in meters",
            "example": 5000
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box: minLon,minLat,maxLon,maxLat",
              "title": "Bbox"
            },
            "description": "Bounding box: minLon,minLat,maxLon,maxLat",
            "example": "34.7,32.0,34.8,32.1"
          },
          {
            "name": "countries",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated ISO-3166-1 alpha-2 codes",
              "title": "Countries"
            },
            "description": "Comma-separated ISO-3166-1 alpha-2 codes",
            "example": "IL,US"
          },
          {
            "name": "categories",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated category filters",
              "title": "Categories"
            },
            "description": "Comma-separated category filters",
            "example": "hotel,restaurant,museum"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "BCP-47 language code",
              "default": "en",
              "title": "Lang"
            },
            "description": "BCP-47 language code",
            "example": "en"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 20,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum results (1-20)",
              "default": 8,
              "title": "Limit"
            },
            "description": "Maximum results (1-20)",
            "example": 8
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of place suggestions with highlighted matching text",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/places/v1/places/search": {
      "get": {
        "tags": [
          "places",
          "places-typeahead",
          "places-typeahead"
        ],
        "summary": "Full Place Search",
        "description": "Search for places with comprehensive details and metadata",
        "operationId": "search_places_places_v1_places_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256,
              "description": "Search query",
              "title": "Q"
            },
            "description": "Search query",
            "example": "hotel montefiore"
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "User latitude for proximity bias",
              "title": "Lat"
            },
            "description": "User latitude for proximity bias",
            "example": 32.07
          },
          {
            "name": "lng",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "User longitude for proximity bias",
              "title": "Lng"
            },
            "description": "User longitude for proximity bias",
            "example": 34.78
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 50000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bias radius in meters",
              "title": "Radius"
            },
            "description": "Bias radius in meters",
            "example": 10000
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box: minLon,minLat,maxLon,maxLat",
              "title": "Bbox"
            },
            "description": "Bounding box: minLon,minLat,maxLon,maxLat",
            "example": "34.7,32.0,34.8,32.1"
          },
          {
            "name": "countries",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated ISO-3166-1 alpha-2 codes",
              "title": "Countries"
            },
            "description": "Comma-separated ISO-3166-1 alpha-2 codes",
            "example": "IL,US"
          },
          {
            "name": "categories",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated category filters",
              "title": "Categories"
            },
            "description": "Comma-separated category filters",
            "example": "hotel,restaurant"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "BCP-47 language code",
              "default": "en",
              "title": "Lang"
            },
            "description": "BCP-47 language code",
            "example": "en"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 50,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum results (1-50)",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum results (1-50)",
            "example": 10
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination token for next page",
              "title": "Page Token"
            },
            "description": "Pagination token for next page",
            "example": "page_20_1234567890"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Result offset for pagination",
              "title": "Offset"
            },
            "description": "Result offset for pagination",
            "example": 0
          },
          {
            "name": "open_now",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter for currently open POIs",
              "title": "Open Now"
            },
            "description": "Filter for currently open POIs",
            "example": true
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SortOrder"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort order",
              "default": "relevance",
              "title": "Sort"
            },
            "description": "Sort order",
            "example": "relevance"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed search results with complete place information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/places/v1/places/{place_id}": {
      "get": {
        "tags": [
          "places",
          "places-typeahead",
          "places-typeahead"
        ],
        "summary": "Place Details",
        "description": "Get comprehensive details for a specific place",
        "operationId": "get_place_details_places_v1_places__place_id__get",
        "parameters": [
          {
            "name": "place_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unique place identifier",
              "title": "Place Id"
            },
            "description": "Unique place identifier",
            "example": "poi_hotel_montefiore"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete place information including contact details, hours, and ratings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/places/v1/places/": {
      "get": {
        "tags": [
          "places",
          "places-typeahead",
          "places-typeahead"
        ],
        "summary": "Places API Health Check",
        "description": "Check the health and status of the Places API service",
        "operationId": "health_check_places_v1_places__get",
        "responses": {
          "200": {
            "description": "Service health status and uptime information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "relevance",
          "distance",
          "rating",
          "popularity"
        ],
        "title": "SortOrder",
        "description": "Search result sorting options"
      },
...
      "SuggestResponse": {
        "properties": {
          "session_token": {
            "type": "string",
            "title": "Session Token",
            "description": "Session token for this response"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/PlaceSuggestion"
            },
            "type": "array",
            "title": "Suggestions",
            "description": "Suggestion results"
          }
        },
        "type": "object",
        "required": [
          "session_token",
          "suggestions"
        ],
        "title": "SuggestResponse",
        "description": "Type-ahead suggestions response"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token in the format: Bearer <token>"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.mytrip.com",
      "description": "Production server"
    }
  ],
...
}